+++ src/MusicVisualizer/Utils/Data/AudioHandling/PortAudioStream.h
@@ -48,8 +48,8 @@ namespace mvlizer::data {
                         std::shared_ptr<IAudioCallback> callbacks);
 
         PaStreamCallbackResult callback(
-                const float* inputBuffer,
-                const float* outputBuffer,
+                const int32_t* inputBuffer,
+                const int32_t* outputBuffer,
                 unsigned long framesPerBuffer,
                 const PaStreamCallbackTimeInfo* timeInfo,
                 PaStreamCallbackFlags flags
